requirements.txt
flask~=3.0.2
c2pa-python
flask_wtf
pillow~=10.2.0
folium~=0.16.0
requests~=2.31.0
Werkzeug~=3.0.1
WTForms~=3.1.2
python-dateutil